clean up + review sql
