+++ docker-compose.yml
@@ -8,8 +8,8 @@ services:
     restart: always
     environment:
       MYSQL_DATABASE: puresynth
-      MYSQL_USER: ${DB_USERNAME}
-      MYSQL_PASSWORD: ${DB_PASSWORD}
+      MYSQL_USER: puresynth
+      MYSQL_PASSWORD: puresynth
       MYSQL_ALLOW_EMPTY_PASSWORD: "yes"
     ports:
       - "3307:3306"
