+++ src/main/java/com/desiato/puresynth/PureSynthApplication.java
@@ -1,6 +1,5 @@
 package com.desiato.puresynth;
 
-import com.desiato.puresynth.repositories.RoleRepository;
 import com.desiato.puresynth.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
@@ -14,9 +13,6 @@ public class PureSynthApplication implements CommandLineRunner {
 
 	@Autowired
 	private UserRepository userRepository;
-
-	@Autowired
-	private RoleRepository roleRepository;
 
 	@Autowired
 	private PasswordEncoder passwordEncoder;
