--- src/main/java/com/desiato/puresynth/controllers/AdminController.java
@@ -1,70 +1,0 @@
-package com.desiato.puresynth.controllers;
-
-import com.desiato.puresynth.models.User;
-import com.desiato.puresynth.services.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/admin")
-public class AdminController {
-
-    @Autowired
-    private UserService userService;
-
-    @GetMapping("/")
-    public String testMapping() {
-        return "Controller is reachable";
-    }
-
-    // Fetch all users
-    @GetMapping
-    public List<User> getAllUsers() {
-        return userService.getAllUsers();
-    }
-
-    // Fetch a user by ID
-    @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id) {
-        return userService.getUserById(id)
-                .map(user -> ResponseEntity.ok(user))
-                .orElse(ResponseEntity.notFound().build());
-    }
-
-    // Create a new user (from the admin side)
-    @PostMapping
-    public ResponseEntity<Object> createUser(@RequestBody User newUser) {
-        if (userService.findByUsername(newUser.getUsername()) != null) {
-            // Username already exists, return a conflict response
-            return ResponseEntity
-                    .status(HttpStatus.CONFLICT)
-                    .body("Username already exists");
-        }
-        User createdUser = userService.saveUser(newUser);
-        return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
-    }
-
-    // Update any user's details
-    @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) {
-        return userService.getUserById(id).map(user -> {
-            user.setUsername(userDetails.getUsername());
-            user.setEmail(userDetails.getEmail());
-            User updatedUser = userService.saveUser(user);
-            return ResponseEntity.ok(updatedUser);
-        }).orElse(ResponseEntity.notFound().build());
-    }
-
-    // Delete any user
-    @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
-        return userService.getUserById(id).map(user -> {
-            userService.deleteUser(id);
-            return ResponseEntity.ok().build();
-        }).orElse(ResponseEntity.notFound().build());
-    }
-}
