--- src/main/java/com/desiato/puresynth/repositories/RoleRepository.java
@@ -1,8 +1,0 @@
-package com.desiato.puresynth.repositories;
-
-import com.desiato.puresynth.models.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findByName(String name);
-}
