+++ src/main/java/com/desiato/puresynth/repositories/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String username);
 
+    User findByEmail(String email);
 }
